chore: rewrite in typescript
I kind of hate myself for doing this
