+++ .mocharc.js
@@ -1,8 +1,9 @@
 'use strict';
 
 module.exports = {
+  require: 'ts-node/register',
   timeout: '2s',
   color: true,
   slow: '1s',
-  'forbid-only': Boolean(process.env.CI)
+  'forbid-only': Boolean(process.env.CI),
 };
