+++ bin/smoker.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require('../dist/cli.js');
