+++ package.json
@@ -23,29 +23,34 @@
   },
   "license": "Apache-2.0",
   "author": "Christopher Hiller <boneskull@boneskull.com> (https://boneskull.com/)",
-  "main": "src/index.js",
-  "types": "types/src/index.d.ts",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "bin": {
-    "smoker": "src/cli.js"
+    "smoker": "./bin/smoker.js"
   },
   "files": [
-    "src",
-    "types/src"
+    "./src",
+    "./dist",
+    "!**/*.tsbuildinfo"
   ],
   "scripts": {
     "build": "tsc -b",
+    "clean": "tsc -b --clean",
+    "dev": "tsc -b --watch",
     "fix": "eslint . --fix",
     "lint": "eslint .",
     "lint:commit": "commitlint",
     "lint:staged": "lint-staged",
-    "prepare": "husky install; npm run build",
-    "test": "mocha \"test/**/*.spec.js\"",
+    "prepare": "husky install; npm run rebuild",
+    "rebuild": "npm run clean; npm run build",
+    "test": "mocha \"test/**/*.spec.ts\"",
     "posttest": "markdownlint-cli2 \"README.md\" && npm run lint",
-    "test:smoke": "node ./src/cli.js --version",
-    "update-snapshots": "cross-env UNEXPECTED_SNAPSHOT=1 npm test"
+    "test:smoke": "node ./bin/smoker.js --version",
+    "update-snapshots": "cross-env UNEXPECTED_SNAPSHOT=1 npm test",
+    "preversion": "npm run rebuild"
   },
   "lint-staged": {
-    "*.js": [
+    "*.{js,ts}": [
       "eslint --fix",
       "prettier --write"
     ],
@@ -53,7 +58,7 @@
       "markdownlint-cli2-fix",
       "prettier --write"
     ],
-    "*.{yml,d.ts}": [
+    "*.yml": [
       "prettier --write"
     ]
   },
@@ -83,6 +88,8 @@
     "@types/sinon": "10.0.16",
     "@types/which": "3.0.0",
     "@types/yargs": "17.0.24",
+    "@typescript-eslint/eslint-plugin": "6.2.0",
+    "@typescript-eslint/parser": "6.2.0",
     "cross-env": "7.0.3",
     "eslint": "8.46.0",
     "eslint-config-prettier": "8.10.0",
@@ -98,6 +105,7 @@
     "prettier": "3.0.1",
     "rewiremock": "3.14.5",
     "sinon": "15.2.0",
+    "ts-node": "10.9.1",
     "typescript": "5.1.6",
     "unexpected": "13.2.1",
     "unexpected-eventemitter": "2.4.0",
