+++ test/stubs.d.ts
